Move python code to root folder
