+++ .gitignore
@@ -136,3 +136,7 @@ dmypy.json
 
 # Cython debug symbols
 cython_debug/
+
+secrets.py
+elmo.db
+.vscode
