--- python/secrets.py
@@ -1,3 +1,0 @@
-STRAVA_CLIENT_ID = 
-STRAVA_CLIENT_SECRET = 
-
