+++ scripts/init_db.py
