+++ secrets.py.example
@@ -1,0 +1,3 @@
+STRAVA_CLIENT_ID = ""
+STRAVA_CLIENT_SECRET = ""
+REDIRECT_URI = ""
